fix (core/scripts) goldshire in elwynn_forrest consolidation
zone_goldshire.cpp = Goldshire is not a zone, it is a area located in Elwynn Forest zone, zone 10.
Consolidated the script into the proper zone and adjusted script loader, declaration, and the call.
